requirements.txt
click==8.1.3
DateTime==4.5
exceptiongroup==1.1.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==23.1
pluggy==1.0.0
PyJWT==2.4.0
pytest==7.3.1
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.41.1
tomli==2.0.1
uuid==1.30
Werkzeug==2.2.2
zipp==3.15.0
zope.interface==5.4.0
